+++ src/contexts/index.ts
@@ -1,6 +1,10 @@
 export { EdgesContext, useEdgesProperties } from './EdgesContext';
 export { GraphContext, useGraphProperties } from './GraphContext';
+export { GraphMetaContext, useGraphMeta } from './GraphMetaContext';
+export { GraphRegistryContext, useGraphRegistry } from './GraphRegistryContext';
+export { GraphSelectionContext, useGraphSelection } from './GraphSelectionContext';
 export { LayoutContext, useLayoutProperties } from './LayoutContext';
+export { ModalsContext, useModals } from './ModalsContext';
 export { NodesContext, useNodeProperties } from './NodesContext';
-export { PopupsContext, useModals, useToasts } from './PopupsContext';
 export { SettingsContext, useSettings } from './SettingsContext';
+export { ToastsContext, useToasts } from './ToastsContext';