+++ src/hooks/useActionBarLogic.ts
@@ -1,9 +1,9 @@
-import { ParsedErrorToast, parseError } from '@/config/parsedError';
+import { ParsedErrorToasts, parseError } from '@/config/parsedError';
 import { useGetGraph } from '@/hooks/useGraphRegistry';
 import { addEdges, removeEdges } from '@/services/edgesService';
+import { resetGraph } from '@/services/graphService';
 import { arrangeGraph, centerGraph } from '@/services/layoutService';
 import { addNode, removeNodes } from '@/services/nodesService';
-import { resetGraph } from '@/services/graphService';
 import { isArrayOfStrings } from '@/types/typeGuards';
 import {
     useGraphProperties,
@@ -55,7 +55,7 @@ export function useActionBarLogic() {
         const graph = graphRef.current;
 
         if (!graph) {
-            addToast(ParsedErrorToast.GraphNotFound);
+            addToast(ParsedErrorToasts.GraphNotFound);
             return;
         }
 
@@ -87,7 +87,7 @@ export function useActionBarLogic() {
     const handleArrangeGraph = useCallback(() => {
         const graph = graphRef.current;
         if (!graph) {
-            addToast(ParsedErrorToast.GraphNotFound);
+            addToast(ParsedErrorToasts.GraphNotFound);
             return;
         }
 
@@ -105,7 +105,7 @@ export function useActionBarLogic() {
     const handleCenterGraph = useCallback(() => {
         const graph = graphRef.current;
         if (!graph) {
-            addToast(ParsedErrorToast.GraphNotFound);
+            addToast(ParsedErrorToasts.GraphNotFound);
             return;
         }
 
@@ -118,7 +118,7 @@ export function useActionBarLogic() {
     const handleAddNode = useCallback(() => {
         const graph = graphRef.current;
         if (!graph) {
-            addToast(ParsedErrorToast.GraphNotFound);
+            addToast(ParsedErrorToasts.GraphNotFound);
             return;
         }
 
@@ -130,7 +130,7 @@ export function useActionBarLogic() {
     const handleAddEdges = useCallback(() => {
         const graph = graphRef.current;
         if (!graph) {
-            addToast(ParsedErrorToast.GraphNotFound);
+            addToast(ParsedErrorToasts.GraphNotFound);
             return;
         }
 
@@ -174,7 +174,7 @@ export function useActionBarLogic() {
 
     const handleDeleteSelected = useCallback(() => {
         if (!graphRef.current) {
-            addToast(ParsedErrorToast.GraphNotFound);
+            addToast(ParsedErrorToasts.GraphNotFound);
             return;
         }