+++ src/main.tsx
@@ -1,8 +1,9 @@
 import { App } from '@/App';
-import { PopupsProvider } from '@/providers/PopupsProvider';
+import { ErrorBoundary } from '@/components/ErrorBoundary';
+import { ModalsProvider } from '@/providers/ModalsProvider';
+import { ToastsProvider } from '@/providers/ToastsProvider';
 import '@/styles/animations.css';
 import '@/styles/main.css';
-import { isDev } from '@/utils/general';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { importCytoscapeExtensions } from './config/extensions';
@@ -16,16 +17,16 @@ if (!rootElement) {
     throw new ParsedError('Root element with id "root" not found');
 }
 
-const appComponent = isDev() ? (
+const appTree = (
     <StrictMode>
-        <PopupsProvider>
-            <App />
-        </PopupsProvider>
+        <ErrorBoundary>
+            <ToastsProvider>
+                <ModalsProvider>
+                    <App />
+                </ModalsProvider>
+            </ToastsProvider>
+        </ErrorBoundary>
     </StrictMode>
-) : (
-    <PopupsProvider>
-        <App />
-    </PopupsProvider>
 );
 
-createRoot(rootElement).render(appComponent);
+createRoot(rootElement).render(appTree);